5초간 자동 api 통신 구현 완성. 리뷰하기, 약속 취소하기, 신고 api 완성. 내일부터 채팅 ui 본격적으로 시작하면 됨

+++ SeSac Service Level Project/HomeViewPart/ChatView/ChattingViewController.swift
@@ -38,20 +38,42 @@ class ChattingViewController: UIViewController {
     }
     
     @objc func reportButtonClicked() {
+        
         let vc = ChatMenuViewController()
         vc.menuType = .report
         vc.mainView.title.text = "새싹 신고"
         vc.mainView.subTitle.text = "다시는 해당 새싹과 매칭되지 않습니다"
         vc.mainView.textView.text = "kiki"
         vc.mainView.confirmButton.setTitle("신고하기", for: .normal)
+        
         vc.modalPresentationStyle = .overFullScreen
         self.present(vc, animated: true, completion: nil)
     }
     @objc func cancelButtonnClicked() {
+        guard let idToken = UserDefaults.standard.string(forKey: "idToken") else { return }
         let vc = DeleteViewController()
         vc.modalPresentationStyle = .overFullScreen
-        vc.mainView.title.text = "약속을 취소하시겠습니까?"
-        vc.mainView.subTitle.text = "약속을 취소하면 패널티가 부과됩니다"
+        vc.mainView.viewType = .dodge
+        if UserInfo.current.dodged == 0 {
+            vc.mainView.subTitle.text = "상대방이 약속을 취소했기 때문에 패널티가 부과되지 않습니다"
+        }
+        vc.uid = UserInfo.current.matchedUid
+        vc.idToken = idToken
+        vc.completion = { statusCode, uid in
+            if statusCode == 200 {
+                HomeApiService.stopFinding(idToken: idToken) { _, statusCode2 in
+                    if statusCode2 == 200 {
+                        UserDefaults.standard.set(0, forKey: "CurrentUserState")
+                        UserInfo.current.matched = 0
+                    }
+                }
+                self.navigationController?.popToRootViewController(animated: true)
+            }else {
+                print(statusCode)
+            }
+        }
+        
+        
         self.present(vc, animated: true, completion: nil)
     }
     @objc func reviewButtonClicked() {
@@ -61,6 +83,9 @@ class ChattingViewController: UIViewController {
         vc.mainView.subTitle.text = "고래밥님과의 취미 활동은 어떠셨나요?"
         vc.mainView.textView.text = "kiki"
         vc.mainView.confirmButton.setTitle("gh", for: .normal)
+        vc.completion = {
+            self.navigationController?.popToRootViewController(animated: true)
+        }
         vc.modalPresentationStyle = .overFullScreen
         self.present(vc, animated: true, completion: nil)
     }
